Move callback tenant detection and authorization to middleware.
