+++ lib/ensureTenantContext.js
@@ -13,10 +13,15 @@ var ensureTenant = {
         var hostParts = req.get('host').split('.');
         // retrieves the tenant name from the host.
         var tenant = hostParts.length > 2 ? hostParts[0] : undefined;
-        var isUserInTenant = req.user._json.groups.some((element, index, array) => {
-          console.log('Index: ' + index);
-          return element.toLowerCase() === tenant.toLowerCase();
-        });
+
+        // if using session and tenant is not in the url check tenant in session.
+        if (req.session && tenant) tenant = tenant || req.session.tenant;
+
+        if (tenant) {
+          var isUserInTenant = req.user._json.groups.some((element, index, array) => {
+            return element.toLowerCase() === tenant.toLowerCase();
+          });
+        }
 
         if (!tenant || !isUserInTenant) {
           if (setReturnTo && req.session) {
